variant method to reload @@anonymous_user User class variable
  User.anonymous(true)

Passing true to this method will force reloading the 
@@anonymous_user User class variable.

This is only needed now when recreating the anonymous 
user with a Factory during testing after clearing the 
database of an earlier Factory creation of an anonymous 
user.

Without doing this the forget_me method restful_authentication
adds to User will throw an error after it clears the 
remember_tokens for the and tries to save to a non-existant 
old anonymous user.

In general we should NOT be caching items like this in class
variables. It's not thread-safe and won't work right when 
running multiple processes.