render_show_partial only removes "Embeddable::"
Previously if an Embeddable::Smartgraph::RangeQuestion
was passing in to this method it would call the partial
and pass in a local variable named: range_question.

Now it will pass in the local variable: smartgraph_range_question

This helper method uses the new core String methods:

  String#delete_module
  String#underscore_module

I wanted the names to be unique in the scope of embeddable 
controllers and views for the meta-programming I
am doing with spec tests for the embeddable controllers.

This has necessitated changes in corresponding view partials.

These commits coming next ...