use AppSettings module to read/write settings
AppSettings is a new module in lib/ that provides methods
for reading and writing all the application settings as
well as reading the RAILS_ENV specific settings.

Three areas in the application use this functionality have 
been refactored to use the new common functions.

Updated in this commit:
  Admin::Project
  load_config initializer

Updated in the next commit:
  config/setup.rb

When settings for a specific RAILS_ENV are loaded the 
hash keys are symbolized.

  AppSettings.load_app_settings

However when loading ALL the settings with this method:

  AppSettings.load_all_app_settings

The top-level keys that match the values returned by
RAILS_ENV are NOT symbolized.

When settings are saved with this method:

  AppSettings.save_app_settings(new_app_settings)

All the keys (beneath the RAILS_ENV key) are first symbolized
and then merged into the existing saved settings (overwriting
duplicates) before writing to the file system.

Once the code is updated and the application started once
the settings saved to the file system with also use symbolized
keys.