User.anonymous creates user if not found
This is a Class method for returning the memoized anonymous user

If you have deleted and recreated the Anonymous user
then call User.anonymous(true) once to reload the memoized
object. If you don't then calling User.anonymous will return
the older deleted Anonymous user.

If the anonymous user can't be found it is created.

FIXME: using class variables like this is not thread-safe