CHANGE rake task names to be more generic
I'm removing the references to the original project from
the names of the rake tasks used to manage the rails-portal.

This task:  rake rigse:setup:new_rites_app
Is now:     rake app:setup:new_app

Most of the rake tasks created specifically for the 
rails-portal are now namespaced to 'app'.

For example to list most of the specific rails-portal tasks:

  rake -T app: