Fix "abstract_class?" bug in the rspec tests
There was a bug we were seeing in rspec tests saying "undefined method `abstract_class?' for Object:Class" when using the dynamic finder "find_by_login".  The code works when it is changed to use "first(:conditions => { :login => ... })", but that would be changing the code to make a bug in tests go away, so I mocked the test object's "find_by_login" method.