get more tests to pass in webdriver IE
part of this is a javascript implementation of "when I add X to the page".  The implementation is tied too closely to the code, but at least the steps themself are more general.  _How_ "an X is added to the page" is up to the step definition, so that same step could be used if the authoring is changed a lot.