move implemenation of assignment into step definition
Besides the refactoring, makes the IE test pass because it uses the scroll_into_view method to make sure the target of the drag is on the screen.  The refactor might pave the way for projects that use a different assignment UI to still run the same steps just with a different implementation.