Add conditional option to use mysql2
This isn't perfect because the Gemfile.lock file will be ping ponging between mysql mysql2.  And the schema.rb file changes as well depending which one is being used.  Bundler has a platforms option, which might help with this, but I didn't see documentation about how to setup a custom platform.