unique activity names breaks investigation duplication
the specs testing this are marked pending so we can get to a green suite.  ITSISU doesn't need to duplicate investigations, but inorder to merge unique activities back to master some work will need to be done.   I think there might be a way to address the "unique name" need without actually making the names be unique.