use master for xproject-dev deploys
This simplifies deploying xproject-dev.  There should be a need to do parallel development on the xproject-dev branch.  If stuff is experimental it should be a topic branch.  And with our better test coverage it is less likely that master will have serious problems even as it is rapidly being developed.