Convert background job migration to explicit spec
The implicit background job migration failed running 
under JRuby because BackgroundJob doesn't work in JRuby.

I've made the migration explicit so it runs under JRuby.

The app won't use the tables when running under JRuby but this
change makes it more practical for a developer to develop
on one instance and switch between different Ruby VMs.

This migration results in identical tables in the database
so there is no need to run it again on an existing project.