config/setup.rb: add --include-otrunk-examples option
Import and introspect content in otrunk-examples, this 
enables the raw_otml embeddable, the default for this option
is false.

If set to true using this option OR if set true in settings.yml
this rake task is run when creating a new app instance:

  rake app:import:generate_otrunk_examples_rails_models

This task must be run either when setting up a new app or
manually after for the otrunk imports to be properly generated
when rendering the otml for a raw_otml embeddable to work.

If you run this task manually you will also need to add this to
the production and development environments in settings.yml:

  :include_otrunk_examples: true