update views for probe models
the views aren't complete but it is enough to add a new vendor interface and device config without going into script console.  Really this ability to manipulate the probe information should be moved external to the the rails-portal, since it should be the same for all portals and would be "authored" by a different set of users.