add GoLink JNA support
This shows up as a second GoLink Vendor Interface with a different device_id.  The JNA shouldn't show up in the UI because it will just confuse users.  To keep out duplicates the portal settings should only include one of the 2 GoLink VendorInterfaces.  We should add a field to VendorInterface to indicate the preferred or default set of VendorInterfaces.