Handle fake ITSI calibrations, fix real calibrations
There are few fake calibrations in the ITSISU DIY, which are just used to set the axis ranges.  These are now handled through a map in the importer code.  Also real calibrations used to be broken in the importer because they data_collector model would incorrectly set the calibration to nil in some cases.