replace OTBlah with real MW model type
the fake OTBlah model type caused problems with the otml generation.  It seems like there would be a better way to handle this templating so that no embeddables need to be created, and instead they are created on the fly as they are needed. But there is no time to do that now.