alternative way to show the user data warning bar
This brings back the no_user mode property and adds a new one user_data_warning.  By taking away the no_user mode it was disabling the "reload" window.  Also I remember better that the explicity no_user mode bypassed some complex logic which was figuring out if learner data was going to be saved or not.  So if there isn't user data it is better to explicitly say no_user.  The new option user_data_warning lets you control if the red bar at the bottom shows up.  It might make sense to use the no_user mode for otrunk authoring steps as well.