use version_str as key for jnlp_object yaml files
The MavenJnlp::VersionedJnlp object uses the Jnlp RubyGem
to load more complete jnlp objects and these objects are
persisted using yaml to a hierarchical directory structure
rooted here : config/jnlp_objects 

This change uses the unique version_str of the jnlp as the key 
when persisting the yaml serialization to the file system 
instead of the primary key of the associated MavenJnlp::VersionedJnlp
object. 

The version_str key will not change for any specific jnlp even if the 
primary key for the associated MavenJnlp::VersionedJnlp does change
through deletion and re-creation.

In addition the suffix of the persisted yaml files has been changed 
from: '.yaml' to '.yml' to follow the convention used in the rest 
of the code.

Here's an example of the persisted jnlp_object yaml files for the
MavenJnlp::MavenJnlpServer server named concord. There are two
families of jnlps persisted: all-otrunk-snapshot and gui-testing.

$ tree config/jnlp_objects/
config/jnlp_objects/
|-- README
`-- concord
    |-- all-otrunk-snapshot
    |   `-- jnlp_object_0.1.0-20110514.160532.yml
    |-- gui-testing
    |   `-- jnlp_object_0.1.0-20110511.203742.yml
    `-- maven_jnlp_server_object.yml