School creation uses Ajax callbacks for description
The main way to create a School now is to select it from a 
list of NCES schools.

When one NCES school is selected (but before the form is saved)
the writable description in the Rich Text editor is updated
with content generated from data in the associated NCES school.