DRY up sail config builders
Now there is really only one xml builder.  The other files just call it with different parameters.  I don't think there are any tests for this config generation.  I didn't see an useful unit test to do of this.  A integrated regression test seems the best.  But we don't have an example of that.  That same style of test would be useful for the otml and jnlp views.  The closest to this is the embeddable controller tests.