use rails session param for config file
previously the session id was passed to the config url with a 'session' key.  That meant that the config response couldn't easily access the session (current user) itself.  With this new approach the config response works like every other response.  If the java code is correctly handling cookies, doing it this way might also eliminate the need for setting up the cookie in the config file.