prefix 'rake' with bundle exec in cap deploy recipes
This app specifies rake 0.8.7, when deploying on the dev
server we do not use a gemset but instead use the most recent
version of rake (which right now is 0.9.0). This causes the
rake tasks to fail when deploying to the dev server.

Starting rake with 'bundle exec' makes sure that rake 0.8.7
is used instead.