Add calibration to show G for magnetic field
By default most Vernier sensors return mT for the units of magnetic field.  So this commit includes a calibration to turn those mT into G.  This commit also has a hack in it.  The units of the normal magnetic field sensor are still set to 'G'.  This is because the digital display picks up the units from the sensor itself, it doesn't have a independent unit setting like the axis labels on the graph.

There is now java code so that data filters can specify precision and units, that would be the preferred way to fix this: the view that generates the data filter should include the units in the form supported by this new java code.