deprecation: WIP: refactor finder_sql invocation
This is a new form for using :finder_sql as part of 
defining an association. This will need to be applied
to all the other models which use :finder_sql.

Rails has deprecated some interactions with ActiveRecord 
association class methods and can print the following 
deprecation:

  String-based interpolation of association conditions 
  is deprecated. Please use a proc instead.

This however this isn't a very helpful clue for the 
problem fixed in this commit.

See: http://blog.locomotivellc.com/post/4058408379/associations-string-interpolations