Fix 'mock_model' and 'stub' methods in Cucumber
This fixes both the existing author_creates_investigation.feature and an 
example sceneria where developer_reads_documentation is run with ln 8
changed to "...resources exist using mocks"

This works both when spork is running and when it is not.

Weirdly, if spork is running, this bug could *also* be fixed by simply adding
World(RSpec::Mocks::ExampleMethods) to the top of env.rb, but this doesn't
work when spork is not running.