Move Enumerable ext out of application.rb and into core_extensions
Tests such as features/researcher_reports.feature were failing because
they couldn't find this method that was supposed to be extended in Enumerable

Can't quite see how this ever worked, because like this Enumberable was being
defined as being inside RailsPortal. Moving it out of RailsPortal fixed the
issue, and core_extensions seemed like a better place for it.