split out view tests from investigations_controller
This way one set of tests can be super simple just focusing on the code in the controller itself.  
The non otml view tests are moved to @wip cucumber tests of investigation duplication
The otml view tests are moved to a new otml specific spec test.  It might make sense to even move that view spec test out into cucumber as well.