Remove cookie_only => false (for now)
Removing this will fix somethings but break others.  It will break the .config requests made by Java.  Those requests are done with a query_parameter including the session_id.  Setting cookie_only => false meant that rails would handle that query_parameter for us.  But it also meant that anyone could access any route with a session_id in the query_parameter.

Besides that lowered security, cookie_only is broken at least in rails 3.0.10.  Sometimes it causes the params array to be setup incorrectly.

We are looking into a more specific way to handle the .config requests instead of a global cookie_only.