workless starts up the worker process automatically
However with this configuration I was getting a new worker process per queued job.  That isn't what is supposed to happen it is supposed to only start up one worker process.

Also, the way workless is listing and killing worker process would fail when multiple apps are run on the same machine.