remove delayed job
delayed job is a very nice way to specify bits of code to run in the background, however the project is very splintered, and I couldn't find a perfect solution for spawning the worker process automatically.  The standard delayed job worker process is started manually.

Given all of this, it still might be the best solution for dealing with running bits of code in the background.  However it seems more important to get the rest of the app working at this point instead of dealing with delayed job.