bundle_logger_spec with and without after_save
it seems there ought be a better a way to do this.  but currently some of the bundle logger spec runs with after_save and enabled and some without it enabled.  There is an interesting issue that with after_save enabled the bundles need to more fully filled out.  However it useful to create partially complete bundle_contents inorder to test the behavior of the bundle_logger code.

So the places in bundle_logger that actually create bundle_contents have after_save turned on to verify they create permissible bundles.  And the places that assume bundle_contents already exist have after_save disabled inorder to mock things out better.