easier way to add embeddables to page
a few parts of the code are still doing embeddable.pages << page, but that doesn't correctly set the user of the page_element.  A better approach for this, it seems to me would be to remove the user from the page_element entirely and have it get the user from its parent object.