example of drying up embeddable controllers
The embeddables are very repetitive and thus hard to maintain.
These 2 examples are pretty simple, but some of them have different behavior.
For example some have new and edit views but many do not.
So continuing this refactoring should deal with those differences, 
most likely we can just remove the differences and have a consistent behavior.

Another 'drying' that can be done is to remove the repetative views
and layouts that each embeddable has.