make jnlp_icon slightly themeable
There is a way to use the theme_support plugin to make images themeable in the theme folders.  But this adds quite a bit of complexity for something simple.  A simple way to handle this would be to make some settings themeable.  Then these themed settings can specify alternative paths for certain images.  I believe there are currently only 2 themed images: the banner and now this jnlp_icon.