Convert non-second x axis limits to seconds in the otml.
The otml expects that axis min/max values are set in seconds, even when the units is set to minutes. Handle this conversion automatically, so that authors can set their limits in minutes and have everything work out.

NOTE: Data in the data store (for students and author data) will still be stored in seconds! No conversions happen on that data.