better error handling when jnlp cannot be found
this includes validating jnlp version strings in the project model
this makes it a little more clear when changing things,
however it might cause some tricky catch 22 when trying
to add a jnlp family.
It still doesn't solve the UX problem when a admin changes the jnlp family in the project
settings form and then can't select a matching version.