[#26145805] printable roster should not show all students in school.
https://www.pivotaltracker.com/story/show/26145805
Printing class roster should not include all students from the school.

Description of bug:
The roster in a class page includes a 'print' link, this leads to a page with only the class roster. However, when this is printed the print out includes a listing of all students in the school.