Add "installer reports".
Basically, when the installer jnlp is used AND opportunistic installers
are turned on, then the installer code will attempt to dynamically download
a jar, run some code from it, and report the results back to the portal.

This will allow us to non-invasively test the feasibility of having the
installer code replace webstart.

Conflicts:

	config/routes.rb
	db/schema.rb