MIGRATION: add jnlp_cdn_hostname to admin_project
if this is set the codebase of the jnlp files is modified to point to this
hostname. This enabled the installer to pull in files from a CDN like
Amazon CloudFront instead of our jnlp server. This can also be used
to serve the files from a simple static website instead of a jnlp server.