fix issue with page ordering
this issue was discovered when duplicating an investigation and the pages were out of
order. The problem could be duplicated just by creating a set of Page objects and
checking their position attribute which is supposed to be managed by acts_as_list.
acts_as_list was recently updated which seems to have changed how it works.
The Page model was setting a default value for the position, this tricked acts_as_list
into thinking that the item had already been added to the list.