include more rails time in bundle upload_time
It is possible that the read method here blocks until all the contents
are uploaded. From testing that doesn't seem the case but I'd expect
in some cases that is how it would work. Also it is probably more
important to use the time for user experience quality, so including
more of the time the user is waiting in this time is a good thing.