anonymizer.js Anonymizer : Teacher can anonymize the teachers report.
https://www.pivotaltracker.com/story/show/47286049
[#47286049]

https://www.pivotaltracker.com/story/show/47006425
[#47006425]

Have a way for teachers to quickly show student data for discussion with other teachers. Can just be an option on the teacher report (click a button and it would scramble the names). Or button to run teacher report has drop-down options that allow 'generate anonymized report' (can be confusing).