Isolate embeddable lookup by embeddable type in process_external_acti…
In LARA we are not sending very unique ID's to the portal (as we should)

Instead we are just sending the type and the id of the obj. in Lara.

Obviously the long range solution would be to make better ID's

But for activities that are already published, scoping to embeddable type seems like a good fallback.