Fix activity publishing to retain embeddables.
[#62332442] On re-publishing an activity from LARA to the portal, the portal was supposed to retain embeddables, but instead it was creating new ones, which orphaned previously created saveables. This fixes the retention of embeddables.

The problem lay in the caching of the embeddables. They were being saved in a temporary hash with the external ID (the LARA ID of the question) as the key. When the cache was checked later, however, the ID was put in as an integer, and the key in the hash was a string. Converting the lookup values to strings before reading from the cache fixed the issue.

https://www.pivotaltracker.com/story/show/62332442