Delayed jobs should ues same Rails env as server
Hardcoding ‘production’ made sense when delayed job was only used in production. However, now, in dev we expect to autospawn delayed jobs. But if delayed job writes to the db specified for the production env, and that is different than the development db, then the delayed jobs won’t happen.