[#95062328][#95625522]
If the district has two records with the same name where one has an lea id and the other has a null lea id, point all records using the null iea district to the district with a valid lea.If a district has two records both with null lea ids, merge into one district.

if one has a null school id, merge it to the one that has a school id. If both have null school ids, merge into one.

For international schools with state code of XX, leave them with a null district and null country so we know to find them and fix them.

Keep track of duplicate users in a model.