remove this unnecessary condition
Note the order of teachers in a class might change because the position
is used for ordering a teacher’s classes. It is not used for ordering
the teachers in a class. So if one teacher moves the class earlier in
their list than the other teacher then list of teachers might flip.