remove handling of 'mine_only' from users index
'mine_only' makes no sense here. If it worked it would show just the current user in the list.
And it turns out it doesn't even work because the user model doesn't have a user_id column anyway,
so the passed in user object would not match anything users table.