update user listing: badges, and hidden buttons
if the current_user is a project_admin they cannot see the teacher or
student pages for users. This is just because there isn’t time to
fix those pages to support project_admins right now.
For consistency now there is a ‘student’ and/or ‘teacher’ badge next
to the user.
This code also fixes an issue with the users#cohorts method previously
it was returning an empty array for students because in ruby [] is true