Center light box when editing. Its hard to find on the screen.
When a teacher is editing a single-page activity and a pop-up box comes up for image or link editing, the light box is hard to find on the screen.

The fix we discussed for this was having the iframe scroll instead of automatically resizing to the size of the page.

[#112067555]
https://www.pivotaltracker.com/story/show/112067555