remove the show page of external activities
This show page had a ‘X’ on it for deleting the activity. It was also
not useful and followed the old style of the portal.

In the process of removing this I had to deal with the ‘return to search’
link on the browse page for external activities.  This link was rendered
regardless of how the user got to the browse page. Now it is only rendered
when they actually come from a search. I renamed the instance variable to
reflect this.

And finally I removed the internal print option for external activities. 
This was broken anyway. It tried to print a page object that didn’t exist.
There is still print support for external activities, through the 
print_url property.

[#136501357]