Optionally send log messages to standard out
This is how logging is handled in LARA on docker. I don’t remember exactly, but it seems
Without this only some initial messages are sent to STDOUT. Without this rails exceptions
and request logs do not appear in the docker logs. We send those logs to AWS Cloudwatch
In production.