automatically redirect user, don’t show error message twice
If the user loads the login page and after_sign_in_path is set and the user is already logged in,
then just redirect to the page. This way a user with multiple tabs that have been logged out
Can just reload them. Also if an external site wants to direct the user to a portal page they
Can construct a URL to this page via the login page and it will automatically take them there
If the user is already logged into portal.

If a user logs in on the auth/login page and there is an after_sign_in_path set, don’t
show them an error message if they don’t have access to this after_sign_in_path page.
Just redirect them to the default sign in page. This way a user won’t see an error message
twice if they are not authorized after logging in.

[#136965555]