Don’t use same database for dev, test, and cucumber
By not setting the DB_NAME the database.yml file will use different database
names for each environment.
In order for Rails to create these databases it needs the DB_USER to be the
root user.
Most of the custom configuration for the mysql service isn’t needed because
Rails is setting up the databases.