Disable CSRF token check on log in
With the CSRF check enabled on log in, a failed CSRF token causes strange behavior.
The session gets cleared by the failed check, but then the login continues as normal.
The result is that the user sees a flash message saying the login was successful, and they are redirected
to the page they would normally see. But the user isn’t actually logged in.
So the page is typically blank and there is a login form in the top right.
[#143357307]