Simplify Selector model by relying on ES
The Selector class now essentially just passes the options to the
ReportLearnersEsController verbatim, and includes an option to get
back a list of up to 5000 learners. (This max could be made settable.)

When we receive back the learner data, we can parse it for a list
of learner IDs, and also for a list of runnables (as type-and-ids).

Given these lists of ids, we can find the set of learners and runnables.

All code related to finding the list of filterables for the filter page
drop-downs is removed, as is the code finding the counts, as the filter
page can do this itself via the API.

[#143600433]