Fix researcher report cucumber test
The parts of the test related to seeing stuff on the page don't work
with the React view. It simulates clicking a button by simply loading
the query url corresponding to the button. The step needs to be changed
to `try to go to` to avoid the validation expectation that occurs with
`am on`, since the query string seems stripped in the validation.

The remaining part of the test still confirms that an excel spreadsheet
is downloading, assuming that the correct response is mocked.