Replace older spreadsheet gem with axlsx gem
There will be failing tests with this commit.
It adds an abstraction ontop of the spreadsheet writing.
That ought to make it easier to test, and it should be easier to switch to a new spreadsheet library.

It also removes the gse parser code. This was requiring the old spreadsheet gem.

[#145550201]