remove the volume from the portal-image test
This volume was added before the solr image was preconfigured with our Sunspot configuration
The use of this volume causes confusion the second time this setup is used for a new image.
The volume from the first time is preserved, so the second time this volume overrides the
the code from the new image.