Merge JWT claims to top-level, add tests
This makes our JWTs match the spec more closely by putting all our
claims at the top level, instead of nested in a `claims` field.

Throw an error if we try and pass in a claim with the same key as one
of our existing reserved claims.

Add tests for JWT claims.