Add learner claims to JWT if grant has a learner
If the bearer access grant passed to the JWT controller has a related
learner, then add the same learner/offering fields to the JWT that are
contained in a lara-style run url. 

Add a JWT controller test, which tests getting a JWT when using the API
both with and without an associated learner.