Bug Fix:  `LearnerActivityFeedback` lookup functions.
Instances where association_ids were being passed in as String values (mostly in tests)
were causing lookup failures.

This commit fixes an assumption that association ids are integers.

[#155408322]

A teacher can select levels of proficiency in rubric-based feedback for each criteria and that score will be saved for that student,.

https://www.pivotaltracker.com/story/show/155408322