Fix:  `api/v1/report_spec.rb` change runnable under test.
`OfferingActivityFeedback` checks if its `runnable` has a `rubric_url`
Created spec to verify `rurbric_feedback` is being used.
Modified spec to accept an `external_activity` as a runnable.

[#155408322]

A teacher can select levels of proficiency in rubric-based feedback for each criteria and that score will be saved for that student,.

https://www.pivotaltracker.com/story/show/155408322