Added peer-to-peer api authentication [#157488572]
This was added to allow LARA to proxy firebase api calls from javascript clients.

This commit also changes the check_for_auth_token() method so that it
now raises an exception when no valid auth token is found instead of
directly rendering an error.

Finally, the new rescue code and the existing recue code that was
listening for Exception have been changed to listen for StandardError
as listening for Exception also swallows errors such as SyntaxError.