Correct spec for proper expectation on user_policy
In master `a_teacher` did not have a cohort so the spec `it { should_not
permit(:confirm) }` was passing. It should have a cohort based on the
test data setup.
https://github.com/concord-consortium/rigse/blob/7c1afc1c1b37719463ea703b790a3e8894e3c4ac/spec/policies/user_policy_spec.rb#L80

After updating to Rspec 2.99 `a_teacher` was properly being created with
a cohort so the spec no longer passed. I changed the spec after
messaging with Scott about what the expected behavior should be.

```
If the active_user is a project admin for project ‘a’. And project ‘a’
has cohort ‘1’, then the active_user should be able to confirm a user in
cohort ‘1’.
```