Remove invalid expect statement
The removed expect statement was returning the expected value,
therefore making the test itself invalid. We spent some time looking
for a way to recreate the intended test but would have required
several stubs. Because a following expect statement in the test
assumes success from the invalid test, we opted to remove it. It was
removed from two tests.