Fix bug causing warning in test output
Turns out that this is the code responsible for warnings like:

```
ignoring attempt to close portal::teacher:0x0000000c9d8558 with html
  opened at byte 6288, line 134
  closed at byte 15429, line 409
  attributes at open: {}
  text around open: "() {\n      return \"#<Portal::Teacher:0x0"
  text around close: "v>\n</div>\n\n\n</body>\n</html>\n"
```

I believe these have been seen in test output since the code was
implemented ~3 years ago.

If I understand correctly, the goal is to detect if the current_visitor
is a teacher or student based on the the value of portal_teacher or
portal_student *not* being nil. What was actually happening is that
to_s was being called on these ruby classes and the result was being
inserted into view.