Remove Image::UseUploadedByInAttribution
I chatted with Noah about this one who was the original author. We agree
that it appears not to be used in practice at this point and it can be
removed.

I cam across this because the constant was redefined in the test suite
which prompted a warning to be displayed.

In practice UseUploadedByInAttribution is never set to true, so the method
`uploaded_by_attribution` would always return an empty string. With that
in mind, I think it is safe to also remove the code that called it.