first attempt to fix broken session
if an auth client in portal doesn't have redirect urls for lara, then after LARA tries to log in the session becomes broken.
If the user tries to log directly into the portal (not going through lara), it doesn't work.
The exception raised shows that the oauth code path is still being traversed.
Unfortunately this fix in this commit doesn't actually fix the problem, but it seems like one piece of it.
I think the reason the problem persists is because something is setting the 'user_return_to' session variable.
And when the exception is raised this session variable remains set. I'm not sure yet where that variable gets set.