remove header login box and code that only it used
the header login box was a partial that was no longer included in the project_header
the main feature it had that isn't supported any more is a hidden popup based login
this was used when a user wanted to log into schoology when the portal was added
as an iframe within schoology. This is an unusual case because usually when the
portal is iframed in schoology the user is automatically logged in.

this code also removes the authCheckPath which used to be sent to portal-pages
in the dynamic_scirpts/_all partial. This authCheckPath was not used by portal pages