simplify redirect on not authorized, update comments
The use of after_sign_in_path here was confusing. If a user tried to access
a page they didn't have access to and they had the after_sign_in_path param
set, then after the access error happened they would be sent to the page pointed
to by that parameter.
It makes more sense that they should be sent to their home page in this case.
In other words this redirect should ignore the after_sign_in_path param.
If this is a report portal then we would not hit this updated redirect call
because there is a conditional checking this beforehand.  So the whole report
portal section of after_sign_in_path_for was ignored.