add test for the existing SSO to schoology and google
Previously the schoology provider was not covered by tests.
This is testing some code that is mainly in gems. However it is nice
to have a test that documents all of the request made from the portal
to these services.
Also it should give us more confidence if when we updated these
gems.

#169231073