support after_sign_in_path param for both google and schoology logins
This required a modified google OAuth2 strategy, and some handling of the
state parameter in the authentications_controller.
The code attempt to be forward compatible for when we upgrade our
OAuth2 gems which will start to include random strings in the state.

I also split the specs into two files and tried to reduce some of the redunancy