Fixed broken country dropdown in teacher registration [#173490184]
The upgrade of react-select changed both the props to the component and how it kept its value which broke the AJAX loading of the countries.

Also some existing jQuery code was being triggered by the update of react-select which caused an error when the dropdown closed.  Code was added to guard against the error.