Fixed cursor jumping in text input [#173795059]
Before this change any update to the text input value would cause the cursor to jump to the end of the line. This was new behavior caused by the update to React and Formsy.

With this change the cursor position is remembered and then restored after the text input's controlled input value is updated in the state.