174373123 —  Add controller and policy specs
This took longer than expected, because there was some untested code.

NEW: It seems that project_admins should not be able to delete projects.
But I discovered that they could.

I made changes to the project_policy now so that project_admins can no longer delete their projects.

Site admins are the only folks to add or remove projects

[#174373123]                https://www.pivotaltracker.com/story/show/174373123