174935633 — view template resolving with themes_on_rails
This fixes issues with missing layouts when using themes_on_rails.

FWIW:  We culd probably remove theme support completely from the portal, and it would make our lives eaier.

The theme attempts to set the layout for all controlelrs based on the theme name. See:

https://github.com/chamnap/themes_on_rails/blob/master/lib/themes_on_rails/action_controller.rb

To change this behavior we set the layout in ApplicationController to 'application'

[#174935633]
https://www.pivotaltracker.com/story/show/174935633